Update 9/17/220
-Fixed Sku Image for Out of stock and Product Alert
-Added Option to delete Deposit on View Modal for Owners
-Fixed Orders Ready to Ship Dropdown Bug

+++ app/Http/Controllers/DepositController.php
@@ -210,11 +210,18 @@ class DepositController extends Controller
         }
         try {
             DB::beginTransaction();
-            $deposit->delete();
-            DB::commit();
-            $output = ['success' => 1,
+            if($deposit->amount <= Auth::user()->available_deposit()) {
+                $deposit->delete();
+                DB::commit();
+                $output = ['success' => 1,
                         'msg' => 'Deposit successfully deleted!'
                     ];
+            }
+            else {
+                $output = ['sucess' => 0,
+                        'msg' => 'Customer available deposit is lower than this deposit amount!'
+                    ];
+            }
         } catch (\Exception $e) {
             \Log::emergency("File:" . $e->getFile(). " Line:" . $e->getLine(). " Message:" . $e->getMessage());
             $output = ['success' => 0,
